examples/slack/requirements.txt
slack-sdk
backoff
python-dotenv
duckyai
fastapi
uvicorn[standard]
groq